+++ packages/zaml-parser/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2018-12-05
+
+### Added
+
+* Add support for inline block.
+
+### Fixed
+
+* `stringify()`: Fix source output of block labels, better block line break.
+
 ## [0.4.1] - 2018-11-30
 
 ### Fixed